pyproject.toml
[tool.poetry]
name = "podder-task-foundation"
version = "0.1.0"
description = ""
authors = ["Takaaki Mizuno <takaaki.mizuno@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^7.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^5.5.4"
yapf = "^0.30.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
